refactor(extensions)!: move FoldingHeading extension from yfm to additional folder
