+++ src/extensions/additional/FoldingHeading/FoldingHeadingSpec/md/skip-heading-sections.ts
