+++ src/extensions/additional/FoldingHeading/commands/openHeadingAndCreateParagraphAfterIfCursorAtEndOfHeading.ts
