+++ src/extensions/additional/FoldingHeading/commands/removeFoldingIfCursorAtStartOfHeading.ts
