+++ src/extensions/additional/FoldingHeading/commands/toggleFoldingOfHeading.ts
