+++ src/extensions/additional/FoldingHeading/const.ts
@@ -1,0 +1,1 @@
+export {YfmHeadingAttr, headingNodeName, type HeadingLevel} from '../../yfm/YfmHeading/const';
