+++ src/extensions/additional/FoldingHeading/index.ts
