+++ src/extensions/additional/FoldingHeading/input-rules.ts
