+++ src/extensions/additional/FoldingHeading/plugins/Folding.ts
