+++ src/extensions/additional/FoldingHeading/plugins/folding.scss
