+++ src/extensions/additional/FoldingHeading/utils.ts
