--- src/extensions/yfm/FoldingHeading/const.ts
@@ -1,1 +1,0 @@
-export {YfmHeadingAttr, headingNodeName, type HeadingLevel} from '../YfmHeading/const';
